Delete Issue

+++ app/api/issues/[id]/route.ts
@@ -31,7 +31,12 @@ export async function PATCH(request: NextRequest, {params}:{params:{id: string}}
 }
 
 export async function DELETE(request: NextRequest, {params}:{params: {id: string}}) {
-    const issue = await prisma.issue.delete({where: {id: parseInt(params.id)}});
+    const issue = await prisma.issue.findUnique({where: {id: parseInt(params.id)}})
+
+    if (!issue)
+        return NextResponse.json('Invalid Issue', {status:400})
+
+    await prisma.issue.delete({where: {id: parseInt(params.id)}});
 
     return NextResponse.json('Issue eliminada.', {status: 200})
 }
